+++ src/pages/feed.tsx
@@ -1,6 +1,6 @@
 import { GetServerSidePropsContext } from 'next';
 import { getServerSession } from 'next-auth/next';
-import { authOptions } from '@/pages/api/auth/options';
+import { authOptions } from '@/pages/api/web/auth/options';
 
 export const getServerSideProps = async ({ req, res }: GetServerSidePropsContext) => {
   const session = await getServerSession(req, res, authOptions);
